+++ src/test/java/com/dir/inno/normalizacion_pdfs/LectorPDFImpreso47VTest.java
@@ -7,7 +7,6 @@ package com.dir.inno.normalizacion_pdfs;
 
 import com.snowtide.PDF;
 import com.snowtide.pdf.Document;
-import com.snowtide.pdf.OutputTarget;
 import com.snowtide.pdf.VisualOutputTarget;
 import java.io.File;
 import java.io.IOException;
@@ -257,6 +256,179 @@ public class LectorPDFImpreso47VTest {
         expResult.get(2).add("12345678");
         expResult.get(2).add("CargoAsignado");
         result = instance6.obtenerNomina();
+        assertEquals(expResult, result);
+    }
+    
+    /**
+     * Test of obtenerDomicilioConst method, of class LectorPDFImpreso47V.
+     */
+    @Test
+    public void testObtenerDomicilioConst() {
+        System.out.println("obtenerDomicilioConst");
+
+        String expResult = "CalleDomicilioConstituido nºcalle, Piso 1, Depto 2,\n"
+                + "Localidaddomconstituido, Departamentodomicilioconst, Entre Ríos\n"
+                + "CP codPost\n"
+                + "Tel.: 3414567897\n"
+                + "E-mail: domicilioConstituido@mail.com";
+        String result = instance1.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+
+        expResult = "Calle constituida 1243, Piso 21, Depto 1,\n"
+                + "Bella Italia, Castellanos, Santa Fe\n"
+                + "CP 2401\n"
+                + "Tel.: 3421123456\n"
+                + "E-mail: email.dominio@algo.com";
+        result = instance2.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+
+        expResult = "dassda asd, Piso 32, Depto 21,\n"
+                + "Bouquet, Belgrano, Santa Fe\n"
+                + "CP 2523\n"
+                + "Tel.: 123321231\n"
+                + "E-mail: 123";
+        result = instance3.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle constituida 1243, Piso 21, Depto 1,\n"
+                + "Bella Italia, Castellanos, Santa Fe\n"
+                + "CP 2401\n"
+                + "Tel.: 3421123456\n"
+                + "E-mail: email.dominio@algo.com";
+        result = instance4.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle 8 333, Piso 05, Depto null,\n"
+                + "Ing. Chanourdie, General Obligado, Santa Fe\n"
+                + "CP 3575\n"
+                + "Tel.: 35435458\n"
+                + "E-mail: lalalal@lelele.lulu";
+        result = instance5.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+        
+        expResult = "CalleDomicilioConstituido nºcalle, Piso null, Depto 2,\n"
+                + "Localidaddomconstituido, Departamentodomicilioconst, Entre Ríos\n"
+                + "CP codPost\n"
+                + "Tel.: 3414567897\n"
+                + "E-mail: domicilioConstituido@mail.com";
+        result = instance6.obtenerDomicilioConst();
+        assertEquals(expResult, result);
+    }
+
+    /**
+     * Test of obtenerDomicilioLegal method, of class LectorPDFImpreso47V.
+     */
+    @Test
+    public void testObtenerDomicilioLegal() {
+        System.out.println("obtenerDomicilioLegal");
+
+        String expResult = "CalleDomicilioLegal nºcalle, Piso 1, Depto 2,\n"
+                + "Localidaddomlegal, Dptodomlegal, Entre Ríos\n"
+                + "CP codPost\n"
+                + "Tel.: 3414567897\n"
+                + "E-mail: domiciliolegal@mail.com";
+        String result = instance1.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle sinnombre 1231, Piso 1, Depto 1,\n"
+                + "Helvecia, Garay, Santa Fe\n"
+                + "CP 3003\n"
+                + "Tel.: 343325216\n"
+                + "E-mail: email@fominio.com";
+        result = instance2.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+
+        expResult = "LLERENA 780, Piso 5, Depto 4,\n"
+                + "Bauer Y Sigel, Castellanos, Santa Fe\n"
+                + "CP 2311\n"
+                + "Tel.: 213231\n"
+                + "E-mail: 231asddsa";
+        result = instance3.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle sinnombre 1231, Piso 1, Depto 1,\n"
+                + "Helvecia, Garay, Santa Fe\n"
+                + "CP 3003\n"
+                + "Tel.: 343325216\n"
+                + "E-mail: email@fominio.com";
+        result = instance4.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle 5656 3245, Piso 4, Depto null,\n"
+                + "Carreras, General Lopez, Santa Fe\n"
+                + "CP 2729\n"
+                + "Tel.: 35535344\n"
+                + "E-mail: vampires@yahoo.ahre";
+        result = instance5.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+                
+        expResult = "CalleDomicilioLegal nºcalle, Piso null, Depto 2,\n"
+                + "Localidaddomlegal, Dptodomlegal, Entre Ríos\n"
+                + "CP codPost\n"
+                + "Tel.: 3414567897\n"
+                + "E-mail: domiciliolegal@mail.com";
+        result = instance6.obtenerDomicilioLegal();
+        assertEquals(expResult, result);
+    }
+    
+    /**
+     * Test of obtenerDomicilioReal method, of class LectorPDFImpreso47V.
+     */
+    @Test
+    public void testObtenerDomicilioReal() {
+        System.out.println("obtenerDomicilioReal");
+        String expResult = "CalleDomRealPlanta numDomRealPlant, Piso null, Depto null,\n"
+                + "Rosario, Rosario, Santa Fe\n"
+                + "CP 2000\n"
+                + "Tel.: Telefono/Fax\n"
+                + "Zonificacion: Parque Industrial\n"
+                + "E-mail: emaildomrealplanta@mail.com";
+        String result = instance1.obtenerDomicilioReal();
+        assertEquals(expResult, result);
+
+        expResult = "Calle ubicacion de planta 09877, Piso 11, Depto 2,\n"
+                + "Aurelia, Castellanos, Santa Fe\n"
+                + "CP 2301\n"
+                + "Tel.: 342827252\n"
+                + "Zonificacion: Industrial/Rural\n"
+                + "E-mail: email.com@dominio.com";
+        result = instance2.obtenerDomicilioReal();
+        assertEquals(expResult, result);
+
+        expResult = "dssadsaads 131212, Piso 21, Depto 2,\n"
+                + "Arequito, Caseros, Santa Fe\n"
+                + "CP 2183\n"
+                + "Tel.: das12321321321\n"
+                + "Zonificacion: Parque Industrial\n"
+                + "E-mail: dsadassda";
+        result = instance3.obtenerDomicilioReal();
+        assertEquals(expResult, result);
+        
+        expResult = "Calle ubicacion de planta 09877, Piso 11, Depto 2,\n"
+                + "Aurelia, Castellanos, Santa Fe\n"
+                + "CP 2301\n"
+                + "Tel.: 342827252\n"
+                + "Zonificacion: Industrial/Rural\n"
+                + "E-mail: email.com@dominio.com";
+        result = instance4.obtenerDomicilioReal();
+        assertEquals(expResult, result);
+        
+        expResult = "Callecinha 89498, Piso 3, Depto null,\n"
+                + "Santa Rosa De Calchines, Garay, Santa Fe\n"
+                + "CP 3022\n"
+                + "Tel.: 668676767\n"
+                + "Zonificacion: Otras Zonas\n"
+                + "E-mail: sadanosdn@ogrio.ogr";
+        result = instance5.obtenerDomicilioReal();
+        assertEquals(expResult, result);
+        
+        expResult = "CalleDomRealPlanta numDomRealPlant, Piso null, Depto 2,\n"
+                + "Rosario, Rosario, Santa Fe\n"
+                + "CP 2000\n"
+                + "Tel.: Telefono/Fax\n"
+                + "Zonificacion: Parque Industrial\n"
+                + "E-mail: emaildomrealplanta@mail.com";
+        result = instance6.obtenerDomicilioReal();
         assertEquals(expResult, result);
     }
 }
